@@ -1,43 +1,37 @@
-namespace PassMeta.DesktopApp.Common.Abstractions
+using System.Collections.Generic;
+using System.Net;
+using PassMeta.DesktopApp.Common.Models.Entities;
+
+namespace PassMeta.DesktopApp.Common.Abstractions;
+
+/// <summary>
+/// Application context: cookies, user, etc.
+/// </summary>
+public interface IAppContext
 {
-    using System.Collections.Generic;
-    using System.Net;
-    using Models.Entities;
+    /// <summary>
+    /// Cookies from the server.
+    /// </summary>
+    public IReadOnlyList<Cookie> Cookies { get; }
 
     /// <summary>
-    /// Application context: cookies, user, etc.
+    /// Application user.
     /// </summary>
-    public interface IAppContext
-    {
-        /// <summary>
-        /// Cookies from the server.
-        /// </summary>
-        public List<Cookie> Cookies { get; set; }
+    public User? User { get; }
 
-        /// <summary>
-        /// Application user.
-        /// </summary>
-        public User? User { get; set; }
+    /// <summary>
+    /// Total count of locally created passfiles.
+    /// </summary>
+    public uint PassFilesCounter { get; }
 
-        /// <summary>
-        /// Total count of locally created passfiles.
-        /// </summary>
-        public uint PassFilesCounter { get; set; }
+    /// <summary>
+    /// Server identifier.
+    /// </summary>
+    public string? ServerId { get; }
 
-        /// <summary>
-        /// Server identifier.
-        /// </summary>
-        public string? ServerId { get; }
-
-        /// <summary>
-        /// Server version. If not null, indicates correct <see cref="IAppConfig.ServerUrl"/>
-        /// and internet connection has been established at least once.
-        /// </summary>
-        public string? ServerVersion { get; }
-
-        /// <summary>
-        /// <see cref="Cookies"/> in form of <see cref="System.Net.CookieContainer"/>.
-        /// </summary>
-        public CookieContainer CookieContainer { get; }
-    }
+    /// <summary>
+    /// Server version. If not null, indicates correct <see cref="IAppConfig.ServerUrl"/>
+    /// and internet connection has been established at least once.
+    /// </summary>
+    public string? ServerVersion { get; }
 }
