@@ -2,7 +2,6 @@ namespace PassMeta.DesktopApp.Common.Abstractions.Services
 {
     using System.Threading.Tasks;
     using PassMeta.DesktopApp.Common.Models.Dto.Request;
-    using PassMeta.DesktopApp.Common.Models.Entities;
 
     /// <summary>
     /// Service for working with user account data.
@@ -10,9 +9,9 @@ namespace PassMeta.DesktopApp.Common.Abstractions.Services
     public interface IAccountService
     {
         /// <summary>
-        /// Get information about current authorized user.
+        /// Refresh information about current authorized user.
         /// </summary>
-        public Task<IResult<User>> GetUserDataAsync();
+        public Task<IResult> RefreshUserDataAsync();
         
         /// <summary>
         /// Edit information about current authorized user.
