@@ -14,12 +14,12 @@ namespace PassMeta.DesktopApp.Common.Abstractions.Services.PassFile
         /// <summary>
         /// Load passfile with encrypted data from the server.
         /// </summary>
-        Task<IResult<PassFile>> GetAsync(int passFileId);
+        Task<IResult<PassFile>> GetInfoAsync(int passFileId);
 
         /// <summary>
         /// Load passfile encrypted data.
         /// </summary>
-        Task<IResult<byte[]>> GetDataAsync(int passFileId, int version);
+        Task<byte[]?> GetDataAsync(int passFileId, int version);
 
         /// <summary>
         /// Load user's passfile list of specified type without encrypted data.
