@@ -2,7 +2,7 @@ namespace PassMeta.DesktopApp.Common.Models
 {
     using System;
     using System.Collections.Generic;
-    using Abstractions.Mapping;
+    using Abstractions.Utils.Mapping;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Linq;
     
@@ -27,7 +27,7 @@ namespace PassMeta.DesktopApp.Common.Models
         /// If not <see cref="Success"/>, short reason.
         /// </summary>
         [JsonProperty("what")]
-        public string? What { get; private set; }
+        public string? What { get; set; }
         
         /// <summary>
         /// Sub-responses. Sub-error information list.
