@@ -1,21 +1,23 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reactive.Linq;
+using System.Threading.Tasks;
+using ReactiveUI;
+    
+using PassMeta.DesktopApp.Common;
+using PassMeta.DesktopApp.Common.Abstractions.Utils.PassMetaClient;
+using PassMeta.DesktopApp.Common.Models.Dto.Response;
+    
+using PassMeta.DesktopApp.Core;
+using AppContext = PassMeta.DesktopApp.Core.AppContext;
+
+using PassMeta.DesktopApp.Ui.ViewModels.Base;
+using PassMeta.DesktopApp.Ui.ViewModels.Logs.Models;
+using PassMeta.DesktopApp.Ui.ViewModels.Journal.Models;
+
 namespace PassMeta.DesktopApp.Ui.ViewModels.Journal
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Reactive.Linq;
-    using System.Threading.Tasks;
-    using Base;
-    using Common;
-    using Common.Models.Dto.Response;
-    using Core.Utils;
-    using Logs.Models;
-    using Models;
-    using ReactiveUI;
-    using Views.Main;
-
-    using AppContext = Core.AppContext;
-
     public class JournalViewModel : PageViewModel
     {
         private static int _pageSize = 50;
@@ -66,6 +68,8 @@ namespace PassMeta.DesktopApp.Ui.ViewModels.Journal
             set => this.RaiseAndSetIfChanged(ref _records, value);
         }
 
+        private IPassMetaClient PassMetaClient => EnvironmentContainer.Resolve<IPassMetaClient>();
+
         public JournalViewModel(IScreen hostScreen) : base(hostScreen)
         {
             LogInfo.RefreshStatics();
@@ -88,7 +92,6 @@ namespace PassMeta.DesktopApp.Ui.ViewModels.Journal
         {
             if (AppContext.Current.User is null)
             {
-                FakeNavigated();
                 TryNavigateTo<AuthRequiredViewModel>(typeof(JournalViewModel));
             }
             else
@@ -101,15 +104,18 @@ namespace PassMeta.DesktopApp.Ui.ViewModels.Journal
 
         private async Task LoadRecordsAsync(int pageIndex)
         {
-            using var preloader = MainWindow.Current!.StartPreloader();
+            using var preloader = AppLoading.General.Begin();
 
             if (pageIndex < 0) return;
 
-            var kind = SelectedKind?.Id > 0 ? "&kind=" + SelectedKind.Id : string.Empty;
-            var pageSize = "page_size=" + _pageSize;
-            var month = "month=" + _selectedMonth.Date.ToString("O");
+            var url = PassMetaApi.History.GetList(_selectedMonth.Date, _pageSize, pageIndex, SelectedKind?.Id > 0 
+                ? new[] { SelectedKind.Id }
+                : null);
 
-            var response = await PassMetaClient.GetAsync<PageResult<JournalRecord>>($"history/pages/{pageIndex}?{month}&{pageSize}{kind}", true);
+            var response = await PassMetaClient.Begin(url)
+                .WithBadHandling()
+                .ExecuteAsync<PageResult<JournalRecord>>();
+
             if (response?.Success is not true) return;
 
             _pageSize = response.Data!.PageSize;
@@ -121,7 +127,7 @@ namespace PassMeta.DesktopApp.Ui.ViewModels.Journal
             }
 
             PageList = pageList;
-            SelectedPageIndex = response.Data.PageNumber - 1;
+            SelectedPageIndex = response.Data.PageIndex;
 
             Records = response.Data.List.Select(rec => new JournalRecordInfo(rec)).ToList();
         }
@@ -132,9 +138,11 @@ namespace PassMeta.DesktopApp.Ui.ViewModels.Journal
 
             if (!Kinds.Any())
             {
-                using var preloader = MainWindow.Current!.StartPreloader();
+                using var preloader = AppLoading.General.Begin();
                 
-                var kindsResponse = await PassMetaClient.GetAsync<List<JournalRecordKind>>("history/kinds", true);
+                var kindsResponse = await PassMetaClient.Begin(PassMetaApi.History.GetKinds())
+                    .WithBadHandling()
+                    .ExecuteAsync<List<JournalRecordKind>>();
 
                 if (kindsResponse?.Success is not true) return;
 
